requirements.txt
streamlit
streamlit-option-menu
pillow
numpy
